FullExecutionTimeMilliseconds is now also a decimal, and if it would be zero (so smaller than what the StopWatch class can measure) it's swapped with HardwareExecutionTimeMilliseconds
--HG--
branch : dev

+++ Hast.Common/Models/IHardwareExecutionInformation.cs
@@ -20,7 +20,7 @@ namespace Hast.Common.Models
         /// Gets the full (including the communication roundtrip) execution time of the hardware execution, in
         /// milliseconds.
         /// </summary>
-        long FullExecutionTimeMilliseconds { get; }
+        decimal FullExecutionTimeMilliseconds { get; }
 
         /// <summary>
         /// Gets the date and time when the execution started.
