+++ Hast.Communication/Models/HardwareExecutionInformation.cs
@@ -6,7 +6,7 @@ namespace Hast.Communication.Models
     public class HardwareExecutionInformation : IHardwareExecutionInformation
     {
         public decimal HardwareExecutionTimeMilliseconds { get; set; }
-        public long FullExecutionTimeMilliseconds { get; set; }
+        public decimal FullExecutionTimeMilliseconds { get; set; }
         public DateTime StartedUtc { get; set; }
